requirements.txt
colorama==0.4.6
langchain==0.1.19
langchain_experimental==0.0.59
langchain_openai==0.1.6
pandas==2.2.1
pydantic==1.10.12
python-dotenv==1.0.1
chroma-hnswlib==0.7.3
chromadb==0.5.0
tiktoken==0.7.0